Changes - Disable accidental resizing feature - Fix arithmetic overflow in vec2math.hpp - Fix zooming

+++ physics engine/vec2math.hpp
@@ -6,6 +6,6 @@ namespace Vec2Math
 {
     float length(Vec2 v)
     {
-        return (float)sqrt(v.x * v.x + v.y * v.y);
+        return (float)sqrt(((double)v.x * (double)v.x) + ((double)v.y * (double)v.y));
     }
 }
